# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Workflow Admin API
  description: 工作流平台管理后台 REST API
  version: 1.0.0
  contact:
    name: Workflow Platform Team
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: https://api.example.com
    description: 生产环境
...
paths:
...
  /api/admin/auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: 用户登录
      description: |
        管理员登录，返回 JWT Token。
        底层认证由 AuthProvider SPI 驱动：
        - local：本地数据库（默认）
        - remote：远程 OAuth 2.0 / OIDC 服务（Keycloak、Casdoor 等）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============ Workflow API ============
...
  /api/admin/schemas/tables:
    get:
      operationId: listTables
      tags: [Schemas]
      summary: 列出数据库中所有用户表
      parameters:
        - name: dataSource
          in: query
          schema:
            type: string
          description: 数据源名称，默认使用主数据源
      responses:
        '200':
          description: 表名列表
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
...
  /api/datasources/{datasourceId}/tables:
    get:
      operationId: listDatasourceTables
      tags: [Datasources]
      summary: 列出指定数据源的所有用户表
      parameters:
        - name: datasourceId
          in: path
          required: true
          description: 数据源名称（对应 listDatasources 返回的 id）
          schema:
            type: string
      responses:
        '200':
          description: 表名列表
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /api/datasources/{datasourceId}/tables/{tableName}/columns:
    get:
      operationId: listDatasourceTableColumns
      tags: [Datasources]
      summary: 列出指定数据源指定表的所有列元数据
      parameters:
        - name: datasourceId
          in: path
          required: true
          schema:
            type: string
        - name: tableName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 列元数据列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DatasourceColumnInfo'
components:
...
  schemas:
...
    DatasourceColumnInfo:
      type: object
      description: 表列元数据
      properties:
        name:
          type: string
          description: 列名
        columnName:
          type: string
          description: 列名（与 name 同义，兼容前端字段）
        dataType:
          type: string
          description: 数据库类型名
        type:
          type: string
          description: 类型（与 dataType 同义，兼容前端字段）
        comment:
          type: string
          description: 列注释
        description:
          type: string
          description: 列描述（与 comment 同义，兼容前端字段）

    # ============ 通用响应 ============
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: integer
          format: int64
          description: Unix 毫秒时间戳
        status:
          type: integer
        error:
          type: string
        path:
          type: string

    # ============ 枚举类型 ============
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: 用户名
        password:
          type: string
          description: 密码
    LoginResponse:
      type: object
      required: [token, type]
      properties:
        token:
          type: string
          description: JWT 访问令牌
        type:
          type: string
          example: Bearer
          description: 令牌类型
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT